+++ isolating_controller/isolation/isolators/affinity.py
@@ -4,7 +4,7 @@ import logging
 from typing import Optional
 
 from isolating_controller.workload import Workload
-from .base_isolator import Isolator
+from .base import Isolator
 from .. import NextStep
 from ...metric_container.basic_metric import MetricDiff
