+++ isolating_controller/isolation/policies/defensive.py
@@ -2,13 +2,13 @@
 
 import logging
 
-from .base_policy import IsolationPolicy
+from .base import IsolationPolicy
 from .. import ResourceType
-from ..isolators import CacheIsolator, CoreIsolator, IdleIsolator, MemoryIsolator
+from ..isolators import CacheIsolator, IdleIsolator, MemoryIsolator, SchedIsolator
 from ...workload import Workload
 
 
-class DiffCPUPolicy(IsolationPolicy):
+class DefensivePolicy(IsolationPolicy):
     def __init__(self, fg_wl: Workload, bg_wl: Workload) -> None:
         super().__init__(fg_wl, bg_wl)
 
@@ -31,30 +31,26 @@ class DiffCPUPolicy(IsolationPolicy):
 
         resource: ResourceType = self.contentious_resource()
 
-        if resource is ResourceType.CPU:
-            self._cur_isolator = self._isolator_map[CoreIsolator]
-            self._cur_isolator._contentious_resource = ResourceType.CPU
-            logger.info(f'Core Isolation for {self._fg_wl} is started to isolate {ResourceType.CPU.name}s')
-            return True
+        if self._is_core_isolated and self._is_mem_isolated and self._is_llc_isolated:
+            self._clear_flags()
+            logger.debug('****All isolators are applicable for now!****')
 
-        elif not self._is_llc_isolated and resource is ResourceType.CACHE:
+        if not self._is_llc_isolated and resource is ResourceType.CACHE:
             self._cur_isolator = self._isolator_map[CacheIsolator]
             self._is_llc_isolated = True
-            logger.info(f'Cache Isolation for {self._fg_wl} is started to isolate {ResourceType.CACHE.name}s')
+            logger.info(f'Cache Isolation for {self._fg_wl} is started')
             return True
 
         elif not self._is_mem_isolated and resource is ResourceType.MEMORY:
             self._cur_isolator = self._isolator_map[MemoryIsolator]
             self._is_mem_isolated = True
-            logger.info(f'Memory Bandwidth Isolation for {self._fg_wl} is started '
-                        f'to isolate {ResourceType.MEMORY.name} BW')
+            logger.info(f'Memory Bandwidth Isolation for {self._fg_wl} is started')
             return True
 
         elif not self._is_core_isolated and resource is ResourceType.MEMORY:
-            self._cur_isolator = self._isolator_map[CoreIsolator]
+            self._cur_isolator = self._isolator_map[SchedIsolator]
             self._is_core_isolated = True
-            self._cur_isolator._contentious_resource = ResourceType.MEMORY
-            logger.info(f'Core Isolation for {self._fg_wl} is started to isolate {ResourceType.MEMORY.name} BW ')
+            logger.info(f'Core Isolation for {self._fg_wl} is started')
             return True
 
         else:
