+++ isolating_controller/isolation/policies/defensive_cpu.py
@@ -2,13 +2,13 @@
 
 import logging
 
-from .base_policy import IsolationPolicy
+from .base import IsolationPolicy
 from .. import ResourceType
-from ..isolators import CacheIsolator, IdleIsolator, MemoryIsolator, SchedIsolator
+from ..isolators import CacheIsolator, CoreIsolator, IdleIsolator, MemoryIsolator
 from ...workload import Workload
 
 
-class DiffPolicy(IsolationPolicy):
+class DefensiveCPUPolicy(IsolationPolicy):
     def __init__(self, fg_wl: Workload, bg_wl: Workload) -> None:
         super().__init__(fg_wl, bg_wl)
 
@@ -31,26 +31,30 @@ class DiffPolicy(IsolationPolicy):
 
         resource: ResourceType = self.contentious_resource()
 
-        if self._is_core_isolated and self._is_mem_isolated and self._is_llc_isolated:
-            self._clear_flags()
-            logger.debug('****All isolators are applicable for now!****')
+        if resource is ResourceType.CPU:
+            self._cur_isolator = self._isolator_map[CoreIsolator]
+            self._cur_isolator._contentious_resource = ResourceType.CPU
+            logger.info(f'Core Isolation for {self._fg_wl} is started to isolate {ResourceType.CPU.name}s')
+            return True
 
-        if not self._is_llc_isolated and resource is ResourceType.CACHE:
+        elif not self._is_llc_isolated and resource is ResourceType.CACHE:
             self._cur_isolator = self._isolator_map[CacheIsolator]
             self._is_llc_isolated = True
-            logger.info(f'Cache Isolation for {self._fg_wl} is started')
+            logger.info(f'Cache Isolation for {self._fg_wl} is started to isolate {ResourceType.CACHE.name}s')
             return True
 
         elif not self._is_mem_isolated and resource is ResourceType.MEMORY:
             self._cur_isolator = self._isolator_map[MemoryIsolator]
             self._is_mem_isolated = True
-            logger.info(f'Memory Bandwidth Isolation for {self._fg_wl} is started')
+            logger.info(f'Memory Bandwidth Isolation for {self._fg_wl} is started '
+                        f'to isolate {ResourceType.MEMORY.name} BW')
             return True
 
         elif not self._is_core_isolated and resource is ResourceType.MEMORY:
-            self._cur_isolator = self._isolator_map[SchedIsolator]
+            self._cur_isolator = self._isolator_map[CoreIsolator]
             self._is_core_isolated = True
-            logger.info(f'Core Isolation for {self._fg_wl} is started')
+            self._cur_isolator._contentious_resource = ResourceType.MEMORY
+            logger.info(f'Core Isolation for {self._fg_wl} is started to isolate {ResourceType.MEMORY.name} BW ')
             return True
 
         else:
